refactor: dir -> root/dir.c
