+++ Makefile.am
@@ -71,6 +71,7 @@ watchman_SOURCES = \
 	main.c       \
 	root.c       \
 	root/ageout.c       \
+	root/dir.c       \
 	root/file.c       \
 	root/iothread.c       \
 	root/lock.c       \
