+++ watchman.h
@@ -764,6 +764,8 @@ bool w_root_init(w_root_t *root, char **errmsg);
 bool remove_root_from_watched(
     w_root_t *root /* don't care about locked state */);
 bool is_vcs_op_in_progress(struct write_locked_watchman_root *lock);
+extern const struct watchman_hash_funcs dirname_hash_funcs;
+void delete_dir(struct watchman_dir *dir);
 
 /* Bob Jenkins' lookup3.c hash function */
 uint32_t w_hash_bytes(const void *key, size_t length, uint32_t initval);
